+++ konva.js
@@ -8,7 +8,7 @@
    * Konva JavaScript Framework v4.0.14
    * http://konvajs.org/
    * Licensed under the MIT
-   * Date: Fri Oct 11 2019
+   * Date: Tue Oct 15 2019
    *
    * Original work Copyright (C) 2011 - 2013 by Eric Rowell (KineticJS)
    * Modified work Copyright (C) 2014 - present by Anton Lavrenov (Konva)
@@ -14261,6 +14261,14 @@
           }
       };
       TextPath.prototype.getSelfRect = function () {
+          if (!this.glyphInfo.length) {
+              return {
+                  x: 0,
+                  y: 0,
+                  width: 0,
+                  height: 0
+              };
+          }
           var points = [];
           this.glyphInfo.forEach(function (info) {
               points.push(info.p0.x);
@@ -14268,10 +14276,10 @@
               points.push(info.p1.x);
               points.push(info.p1.y);
           });
-          var minX = points[0];
-          var maxX = points[0];
-          var minY = points[0];
-          var maxY = points[0];
+          var minX = points[0] || 0;
+          var maxX = points[0] || 0;
+          var minY = points[1] || 0;
+          var maxY = points[1] || 0;
           var x, y;
           for (var i = 0; i < points.length / 2; i++) {
               x = points[i * 2];